requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
